Highlights the sentiment of text.

Takes as input a dataframe from the mine_sentiment function with text and sentiment columns and outputs the highlighted sentences by by sentiment polarity as an html file. This function has been adapted from the highlight function from the sentimentr package by Tyler Rinker.

highlight_sentiment(
  df,
  file = file.path(tempdir(), "polarity.html"),
  digits = 3,
  open = TRUE
)

Arguments

df
A mine_sentiment dataframe containing the text and sentiment columns.

file
The name of the html file output. Default is a temporary file.

digits
The option to split on sentences within the character vector.

open
Logical. If TRUE the text highlighting document will attempt to be opened.

Value

Creates an html document with the sentiment highlighting.

Details

Generates an html document containing the polarity highlighting for each text where positive = green and negative = pink.

See also

Examples

text <- c("I really love working at Plant and Food Research. It is great.",
          "Today I investigated methods in sentiment analysis.",
          "The results were not very pleasing. ðŸ˜’",
          "Sentiment anaylsis is amazing! Especially for emojis. ðŸ’—ðŸ˜€")
mine_sentiment(text = text) %>% highlight_sentiment()

#> Error in mine_sentiment(text = text) %>% highlight_sentiment(): could not find function "%>%"

mine_sentiment(text = text, by = "sentence") %>% highlight_sentiment()

#> Error in mine_sentiment(text = text, by = "sentence") %>% highlight_sentiment(): could not find function "%>%"