Extract the text data from a file

Takes a text file, a file type and two character vectors to specify the range of text data to be extracted and returns the text data between the specified headers as a dataframe. The function taeks docx, doc and pdf files as input.

mine_text(path = "./temp_folder", type = NULL, from = NULL, to = NULL)

Arguments

path
The path to the file or folder containing the files.

type
The type of file to be extracted.

from
The character vector from where the text will be extracted.

to
The character vector to where the text will be extracted.

Value

A dataframe containing character vector(s) of the text data.

Details

This function returns a dataframe of the text from one or multiple files where each row contains the text from one file.

See also

Examples

text_df <- mine_text(type = "docx", path = "./temp_folder", from = "Impact Highlight", to = "Outputs")

#> Error in mine_text(type = "docx", path = "./temp_folder", from = "Impact Highlight",     to = "Outputs"): Please specify the correct path or file name