Process a text.

Process a list of text or character vector by converting to lowercase, replacing contractions, removing numbers and punctuation, and stemming/lemmatising.

process_text(text = NULL, lower = TRUE, contractions = FALSE,
  remove_numbers = TRUE, punctuation = TRUE, remove_words = FALSE,
  lemmatise = FALSE, stem = FALSE)

Arguments

text
The text to be processed.

lower
Convert text to lowercase (TRUE/FALSE).

contractions
Replace contracted words in the text (TRUE/FALSE). If true, 'contractions_dict' will be used.

remove_numbers
Remove numbers in the text (TRUE/FALSE).

punctuation
Remove punctuation in the text (TRUE/FALSE/regex).

remove_words
Remove words in the text (TRUE/FALSE/character vector). If true, 'smart_stopwords' will be used.

lemmatise
Lemmatise the text.

stem
Stem the text.

Value

The processed text as a dataframe

Details

This function is a short, personalised wrapper for many tedious text processing tasks like removing characters or patterns. For more formatting and text processing, see qdap by Tyler Rinker, or tm.

See also

Examples

text <- c("I don't want to process this comment!", "I'm v serious.")
process_text(text, contractions = TRUE)

#> [1] "i dont want to process this comment" "im very serious"                    

process_text(text, contractions = TRUE, punctuation = "[^,.[:^punct:]]+")

#> [1] "i dont want to process this comment" "im very serious."