Scrape a list or table from iPlant

Takes a url to an iPlant webpage and an iPlant handle as input and returns the list data in the form of a tidy dataframe. The webpage should contain a list or a table that the user wishes to scrape.

scrape_list(url = NULL, handle = NULL)

Arguments

url
The url of the iPlant page containing a list or table.

handle
The authenticated iplant handle of the user.

Value

A dataframe containing the iPlant list data.

Details

This function calls scrape_uuid to retrieve the url and UUID of an iPlant page containing a list or table. The request to the page is authenticated with the user's username and password using the iplant handle and the XML data is read in. The handle is either created prior using the iplant_handle function or it is created via scrape_list. The column names are read to a list and each attribute of the XML encoded names is iteratively added to its corresponding cell as either the element value or 'NA' if missing. The list is then converted to a dataframe and any employee ID numbers are removed.

See also

Examples

url <- "https://iplant.plantandfood.co.nz/project/datamgmt/Lists/Science%20Data%20Types/AllItems.aspx"
url <- "https://iplant.plantandfood.co.nz/project/datamgmt/Lists/Data%20Improvement%20Plan/AllItems.aspx"
url <- "https://iplant.plantandfood.co.nz/Team/sp/DataSci/Lists/Data%20Science%20Registration/AllItems.aspx"
df <- scrape_list(url)

if (FALSE) {
url <- "https://iplant.plantandfood.co.nz/datamanagement/_vti_bin/owssvr.dll?XMLDATA=1&List={6981327F-9AFC-42AB-81D3-436BCB93D5A6}&View={A7C7766E-488F-497B-8195-2CC5E40A0526}&RowLimit=0&RootFolder=%2fdatamanagement%2fLists%2fData%20Stewards"
df <- scrape_list(url)
}