Read an Excel spreadsheet directly into R.

Takes a url to an Excel spreadsheet on iPlant, an iPlant handle, and other Excel parameters as input, formats the url then reads the Excel spreadsheet directly in to R.

scrape_xl(url = NULL, handle = NULL, ...)

Arguments

url
The url of the iPlant page containing a Excel spreadsheet.

handle
The authenticated iplant handle of the user.

...
Arguments to be passed to readxl::read_excel, e.g., sheet, range, col_names, etc.

Value

A dataframe containing the Excel data from iPlant.

Details

This function uses an authenticated iPlant handle to securely pass the user's credentials tp request a response to a page containing an Excel spreadsheet on iPlant. It then saves the Excel file extracted from the formatted url to a temporary file and reads the data in to a dataframe. This is an automatic wrapper for 'read_excel' from the readxl pacakge which calls 'excel_format' to determine the path of the Excel file.

See also

Examples

url <- "https://iplant.plantandfood.co.nz/project/I180822/_layouts/15/WopiFrame2.aspx?sourcedoc=/project/I180822/Research/T3_Maj_Diameter_Manual.xlsx&action=default"
url <- "https://iplant.plantandfood.co.nz/project/P345110-01/_layouts/15/WopiFrame.aspx?sourcedoc=/project/P345110-01/Research/VE_fungi_metabarcoding_Y1and2_190121.xlsx&action=default"
my_handle <- iplant_handle()
xl_df <- scrape_xl(url, my_handle)

if (FALSE) {
url <- "https://iplant.plantandfood.co.nz/project/I180822/_layouts/download.aspx?SourceUrl=/project/I180822/Research/T3_Maj_Diameter_Manual.xlsx"
xl_df <- scrape_xl(url)
}